+++ src/classes/HexagonInput.ts
